Use `static_cast`

+++ esphome/components/dht/dht.cpp
@@ -186,15 +186,15 @@ bool HOT IRAM_ATTR DHT::read_sensor_(float *temperature, float *humidity, bool r
     if (checksum_a == data[4]) {
       // Data format: 8bit integral RH data + 8bit decimal RH data + 8bit integral T data + 8bit decimal T data + 8bit
       // check sum - some models always have 0 in the decimal part
-      const uint16_t raw_temperature = uint16_t(data[2]) * 10 + (data[3] & 0x7F);
-      *temperature = raw_temperature / 10.0f;
+      const uint16_t raw_temperature = static_cast<uint16_t>(data[2]) * 10 + (data[3] & 0x7F);
+      *temperature = static_cast<float>(raw_temperature) / 10.0f;
       if ((data[3] & 0x80) != 0) {
         // negative
         *temperature *= -1;
       }
 
-      const uint16_t raw_humidity = uint16_t(data[0]) * 10 + data[1];
-      *humidity = raw_humidity / 10.0f;
+      const uint16_t raw_humidity = static_cast<uint16_t>(data[0]) * 10 + data[1];
+      *humidity = static_cast<float>(raw_humidity) / 10.0f;
     } else {
       // For compatibility with DHT11 models which might only use 2 bytes checksums, only use the data from these two
       // bytes
@@ -219,10 +219,10 @@ bool HOT IRAM_ATTR DHT::read_sensor_(float *temperature, float *humidity, bool r
       return false;
     }
 
-    *humidity = raw_humidity * 0.1f;
+    *humidity = static_cast<float>(raw_humidity) * 0.1f;
     if (*humidity > 100.0f)
       *humidity = NAN;
-    *temperature = int16_t(raw_temperature) * 0.1f;
+    *temperature = static_cast<int16_t>(raw_temperature) * 0.1f;
   }
 
   if (*temperature == 0.0f && (*humidity == 1.0f || *humidity == 2.0f)) {
